Updated imports, added geometry and other things to focus, fixed handler
Moved NSApplicaiton and NSApp to be imported from AppKit
Added Geometry and some other things to Focus, it is needed to get
things to work together with how selfspy works. Right now everything is
hardcoded.
Do not need app var, but need the NSApplication.sharedApplication() to
be called.

Handler now sends things to hooks as it is supposed to

added what to do if called.. for test purposes